refactor(module.d.ts): npm module declaration file to help with ts
i wanted to use evervault with ts but couldnt due to missing types. so, i made a declaration file. i
dont know if it works when you try to import the module upon installation from npm but i run a few
quick tests locally by importing it directly from the fs and it went smoothly